+++ ch10/DesignPatternsSample/FactorySample/ConcreteProduct/ItalianPaymentService.cs
@@ -7,7 +7,7 @@ namespace DesignPatternsSample.FactorySample.ConcreteProduct
     class ItalianPaymentService : IPaymentService
     {
         public string EmailToCharge { get ; set ; }
-        public float MoneyToCharge { get ; set ; }
+        public decimal MoneyToCharge { get ; set ; }
         public EnumChargingOptions OptionToCharge { get ; set ; }
 
         public bool ProcessCharging()
